Multiple stories, sleep/wake, saving/loading, fonts
